Moved all tests into local __tests__ directory
